(REF) CRM_Queue_Queue_Sql* - Extract trait for common functions
Most of the methods in CRM/Queue/Queue/Sql.php and CRM/Queue/Queue/SqlParallel.php are
identical. Move them to a common trait so that they're easier to patch.
